+++ __tests__/benchmark.test.ts
@@ -1,5 +1,4 @@
 describe('@benchmark', () => {
-  it('should pass', () => {
-    expect(1 === 1);
+  it('should pass on normal stream', () => {
   });
 });
